refactor create article functionality
