+++ src/__tests__/reducers/articlesReducer.spec.js
@@ -1,17 +1,17 @@
-import { articleReducer, initialState } from '../../reducers/articlesReducer';
+import { articlesReducer, initialState } from '../../reducers/articlesReducer';
 
 import { GET_ARTICLES_BEGIN, GET_ARTICLES_SUCCESS, GET_ARTICLES_FAIL } from '../../action-types';
 import { articles as articlesMock } from '../../__mocks__/testsMockData/articles.mock-data';
 
 describe('articleReducers', () => {
   test('should return initial state', () => {
-    expect(articleReducer(undefined, {})).toEqual({
+    expect(articlesReducer(undefined, {})).toEqual({
         ...initialState
     });
   });
 
   test('should handle GET_ARTICLES_BEGIN action', () => {
-    expect(articleReducer(initialState, 
+    expect(articlesReducer(initialState, 
         {type: GET_ARTICLES_BEGIN}
     )).toEqual({
         ...initialState,
@@ -20,7 +20,7 @@ describe('articleReducers', () => {
   });
 
   test('should handle GET_ARTICLES_SUCCESS action', () => {
-    expect(articleReducer(initialState, 
+    expect(articlesReducer(initialState, 
         { type: GET_ARTICLES_SUCCESS, payload: articlesMock, count: 1 }
     )).toEqual({
         ...initialState,
@@ -31,7 +31,7 @@ describe('articleReducers', () => {
   });
 
   test('should handle GET_ARTICLES_FAIL action', () => {
-    expect(articleReducer(initialState, 
+    expect(articlesReducer(initialState, 
         { type: GET_ARTICLES_FAIL, payload: 'Request failed with status code 404' }
     )).toEqual({
         ...initialState,
